revert webpack changes
