+++ package.json
@@ -27,7 +27,7 @@
     "create-merged-coverage": "nyc report --reporter=text-lcov > coverage.lcov",
     "report-coverage": "./scripts/codecov -C $COMMIT_HASH_FULL -p app/ -B $BRANCH_NAME -f coverage.lcov",
     "publish-coverage": "yarn normalize-jest-coverage && yarn create-merged-coverage && yarn report-coverage",
-    "webpack": "node --max_old_space_size=2048 -r dotenv/config -r @babel/register node_modules/.bin/webpack --config ./webpack"
+    "webpack": "node --max_old_space_size=2048 -r dotenv/config -r @babel/register node_modules/.bin/webpack --config ./webpack.config.js"
   },
   "resolutions": {
     "babel-core": "^7.0.0-0",
@@ -135,7 +135,6 @@
     "stylus": "0.54.5",
     "superagent": "3.7.0",
     "superagent-bluebird-promise": "3.0.2",
-    "terser-webpack-plugin": "1.3.0",
     "ts-loader": "4.1.0",
     "typescript": "3.0.1",
     "typescript-styled-plugin": "0.1.2",
@@ -212,6 +211,7 @@
     "tslint-config-prettier": "1.15.0",
     "tslint-react": "3.6.0",
     "typeahead.js": "0.10.5",
+    "uglifyjs-webpack-plugin": "1.2.7",
     "webpack": "4.30.0",
     "webpack-cli": "3.2.1",
     "webpack-dev-middleware": "3.5.1",
