+++ scripts/assets.sh
@@ -6,7 +6,7 @@ set -e -x
 
 yarn clean
 
-NODE_ENV=production node_modules/.bin/webpack
+NODE_ENV=production yarn webpack
 stylus \
   $(find src/client/assets -name '*.styl') \
   --compress \
