+++ src/client/components/layout/templates/scripts.jade
@@ -2,5 +2,4 @@
 != sharify.script()
 
 //- Asset package
-script( src=asset('/assets/runtime-manifest.js') )
 script( src=asset('/assets/main.js') )
